Edi client now using v3 routes (#91)

+++ edi-adapter-model/src/main/kotlin/no/nav/helsemelding/ediadapter/model/v3/Notification.kt
@@ -25,5 +25,5 @@ data class Notification(
     val notificationTriggeredByHerId: Int? = null,
     val description: String? = null,
     val createdAt: String? = null,
-    val offset: Long
+    val offset: Int
 )
